Organize Line's vars and separate extensions out to different files
