+++ Sources/SwiftUICharts/LineChart/Line.swift
@@ -9,24 +9,30 @@
 import SwiftUI
 
 public struct Line: View {
+    
     @ObservedObject var data: ChartData
+    
+    var gradient: GradientColor = GradientColor(start: Colors.GradientPurple, end: Colors.GradientNeonBlue)
+    var index:Int = 0
+    let padding:CGFloat = 30
+    var curvedLines: Bool = true
+    
+    @State var showBackground: Bool = true
+    @State private var showFull: Bool = false
+    
     @Binding var frame: CGRect
     @Binding var touchLocation: CGPoint
     @Binding var showIndicator: Bool
     @Binding var minDataValue: Double?
     @Binding var maxDataValue: Double?
-    @State private var showFull: Bool = false
-    @State var showBackground: Bool = true
-    var gradient: GradientColor = GradientColor(start: Colors.GradientPurple, end: Colors.GradientNeonBlue)
-    var index:Int = 0
-    let padding:CGFloat = 30
-    var curvedLines: Bool = true
+    
     var stepWidth: CGFloat {
         if data.points.count < 2 {
             return 0
         }
         return frame.size.width / CGFloat(data.points.count-1)
     }
+    
     var stepHeight: CGFloat {
         var min: Double?
         var max: Double?
@@ -49,17 +55,33 @@ public struct Line: View {
         }
         return 0
     }
+    
     var path: Path {
         let points = self.data.onlyPoints()
-        return curvedLines ? Path.quadCurvedPathWithPoints(points: points, step: CGPoint(x: stepWidth, y: stepHeight), globalOffset: minDataValue) : Path.linePathWithPoints(points: points, step: CGPoint(x: stepWidth, y: stepHeight))
+        if curvedLines {
+            return Path.quadCurvedPathWithPoints(points: points, step: CGPoint(x: stepWidth, y: stepHeight), globalOffset: minDataValue)
+        } else {
+            return Path.linePathWithPoints(points: points, step: CGPoint(x: stepWidth, y: stepHeight))
+        }
     }
+    
     var closedPath: Path {
         let points = self.data.onlyPoints()
-        return curvedLines ? Path.quadClosedCurvedPathWithPoints(points: points, step: CGPoint(x: stepWidth, y: stepHeight), globalOffset: minDataValue) : Path.closedLinePathWithPoints(points: points, step: CGPoint(x: stepWidth, y: stepHeight))
+        if curvedLines {
+            return Path.quadClosedCurvedPathWithPoints(points: points, step: CGPoint(x: stepWidth, y: stepHeight), globalOffset: minDataValue)
+        } else {
+            return Path.closedLinePathWithPoints(points: points, step: CGPoint(x: stepWidth, y: stepHeight))
+        }
+    }
+    
+    func getClosestPointOnPath(touchLocation: CGPoint) -> CGPoint {
+        let closest = self.path.point(to: touchLocation.x)
+        return closest
     }
     
     public var body: some View {
         ZStack {
+            
             if(self.showFull && self.showBackground){
                 self.closedPath
                     .fill(LinearGradient(gradient: Gradient(colors: [Colors.GradientUpperBlue, .white]), startPoint: .bottom, endPoint: .top))
@@ -68,6 +90,7 @@ public struct Line: View {
                     .transition(.opacity)
                     .animation(.easeIn(duration: 1.6))
             }
+            
             self.path
                 .trim(from: 0, to: self.showFull ? 1:0)
                 .stroke(LinearGradient(gradient: gradient.getGradient(), startPoint: .leading, endPoint: .trailing) ,style: StrokeStyle(lineWidth: 3, lineJoin: .round))
@@ -76,23 +99,20 @@ public struct Line: View {
                 .animation(Animation.easeOut(duration: 1.2).delay(Double(self.index)*0.4))
                 .onAppear {
                     self.showFull = true
-            }
-            .onDisappear {
-                self.showFull = false
-            }
-            .drawingGroup()
+                }
+                .onDisappear {
+                    self.showFull = false
+                }
+                .drawingGroup()
+            
             if(self.showIndicator) {
                 IndicatorPoint()
                     .position(self.getClosestPointOnPath(touchLocation: self.touchLocation))
                     .rotationEffect(.degrees(180), anchor: .center)
                     .rotation3DEffect(.degrees(180), axis: (x: 0, y: 1, z: 0))
             }
+            
         }
-    }
-    
-    func getClosestPointOnPath(touchLocation: CGPoint) -> CGPoint {
-        let closest = self.path.point(to: touchLocation.x)
-        return closest
     }
     
 }
